#xxxxx DTXMania Rev:0c282febd430a19882761187d8704f9829cc36c7の修正を適用 #xxxxx DTXCreatorのapp.configを微修正

+++ FDK17プロジェクト/コード/00.共通/CWin32.cs
@@ -342,23 +342,23 @@ namespace FDK
 		[return: MarshalAs( UnmanagedType.Bool )]
 		[DllImport( "user32.dll", CharSet = CharSet.Auto )]
 		public static extern bool IsZoomed( IntPtr hWnd );
-		[DllImport( "winmm.dll" )]
-		public static extern uint midiInClose( uint hMidiIn );
-		[DllImport( "winmm.dll" )]
-		public static extern uint midiInGetDevCaps( uint uDeviceID, ref MIDIINCAPS lpMidiInCaps, uint cbMidiInCaps );
-		[DllImport( "winmm.dll" )]
-		public static extern uint midiInGetID( uint hMidiIn, ref uint puDeviceID );
-		[DllImport( "winmm.dll" )]
-		public static extern uint midiInGetNumDevs();
-		[DllImport( "winmm.dll" )]
-		public static extern uint midiInOpen( ref uint phMidiIn, uint uDeviceID, MidiInProc dwCallback, int dwInstance, int fdwOpen );
-		[DllImport( "winmm.dll" )]
-		public static extern uint midiInReset( uint hMidiIn );
-		[DllImport( "winmm.dll" )]
-		public static extern uint midiInStart( uint hMidiIn );
-		[DllImport( "winmm.dll" )]
-		public static extern uint midiInStop( uint hMidiIn );
-		[DllImport( "user32.dll", CharSet = CharSet.Auto )]
+        [DllImport( "winmm.dll" )]
+        public static extern uint midiInClose( IntPtr hMidiIn );
+        [DllImport( "winmm.dll" )]
+        public static extern uint midiInGetDevCaps( uint uDeviceID, ref MIDIINCAPS lpMidiInCaps, uint cbMidiInCaps );
+        [DllImport( "winmm.dll" )]
+        public static extern uint midiInGetID( IntPtr hMidiIn, ref IntPtr puDeviceID );
+        [DllImport( "winmm.dll" )]
+        public static extern uint midiInGetNumDevs();
+        [DllImport( "winmm.dll" )]
+        public static extern uint midiInOpen( ref IntPtr phMidiIn, uint uDeviceID, MidiInProc dwCallback, IntPtr dwInstance, int fdwOpen );
+        [DllImport( "winmm.dll" )]
+        public static extern uint midiInReset( IntPtr hMidiIn );
+        [DllImport( "winmm.dll" )]
+        public static extern uint midiInStart( IntPtr hMidiIn );
+        [DllImport( "winmm.dll" )]
+        public static extern uint midiInStop( IntPtr hMidiIn );
+        [DllImport( "user32.dll", CharSet = CharSet.Auto )]
 		public static extern IntPtr MonitorFromWindow( IntPtr hwnd, uint dwFlags );
 		[return: MarshalAs( UnmanagedType.Bool )]
 		[DllImport( "user32.dll", CharSet = CharSet.Auto )]
@@ -806,8 +806,8 @@ namespace FDK
 		}
 
 
-		// Win32 メッセージ処理デリゲート
+        // Win32 メッセージ処理デリゲート
 
-		public delegate void MidiInProc( uint hMidiIn, uint wMsg, int dwInstance, int dwParam1, int dwParam2 );
-	}
+        public delegate void MidiInProc( IntPtr hMidiIn, uint wMsg, int dwInstance, int dwParam1, int dwParam2 );
+    }
 }
